requirements.txt
fastmcp>=2.0.0
simple-salesforce>=1.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
cryptography>=42.0.0
